facerecon_backend/facerecon/requirements.txt
aiosqlite==0.16.1
asyncpg==0.22.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
iso8601==0.1.14
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
PyPika==0.44.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
